Fix a quadratic-name hang and two bounds bugs in BombShell/REZ/Toshi parsers
- BombShell: bs_unique() retried _2, _3, ... with a full rescan per try,
  cubic in the asset count; a pack repeating one record BS_MAX_ASSETS
  times spun for minutes. One pass now finds the highest suffix in use.
- REZ: a skeleton that failed rz_bones() validation kept o->bones and
  o->nbones, and ParseRezModel() (which keeps a failed walk's nodes) then
  walked it with rz_joints(), reading and writing out of bounds.
- Toshi: ts_ptr() took a u32 length, truncating the u64 count * stride
  products OpenToshiTkl() checks, so a huge keyframe table passed.

+++ project/src/lib-toshi.c
@@ -159,7 +159,12 @@ s64 ToshiSymbol (const toshi_trb_t *t, ccp name)
 	return -1;
 }
 
-static bool ts_ptr (const toshi_trb_t *t, u32 off, u32 len) { return off <= t->sect_size && len <= t->sect_size - off; }
+// 64-bit on purpose: callers pass count * stride products that must not be
+// truncated before the check.
+static bool ts_ptr (const toshi_trb_t *t, u64 off, u64 len)
+{
+	return off <= t->sect_size && len <= t->sect_size - off;
+}
 
 static uint ts_ttl_entries (const toshi_trb_t *t, u32 *ents)
 {
